# openapi.yaml
openapi: 3.1.0
info:
  title: ConstructInvoice AI API
  description: |
    AI-powered G702/G703 construction billing platform for General Contractors and Subcontractors.
    Manage projects, create pay applications, track payments, and get AI-powered billing intelligence.
  version: "3.0"
  contact:
    name: ConstructInvoice AI Support
    email: support@varshyl.com
  license:
    name: Proprietary
servers:
  - url: https://constructinv.varshyl.com
    description: Production
  - url: https://construction-ai-billing-staging.up.railway.app
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
security:
  - BearerAuth: []
paths:
...
  /api/ai/ask:
    post:
      operationId: askBillingAI
      summary: Ask construction billing AI questions
      tags:
        - AI Assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  description: Question about billing, G702/G703 forms, retainage, liens, cash flow, etc.
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  response:
                    type: string
                  sources:
                    type: array
                    items:
                      type: string
        '400':
          description: Invalid question
...